fix ui

+++ src/main/java/com/poly/polystore/core/client/service/BaoHanhService.java
@@ -7,5 +7,5 @@ import org.springframework.data.domain.Pageable;
 import java.util.List;
 
 public interface BaoHanhService {
-    List<WarrantyDTO> thongTinBaoHanh(String imei, Long idKH, String phoneNumber);
+    List<WarrantyDTO> thongTinBaoHanh(String imei, Integer idKH, String phoneNumber);
 }
